Выделен абстрактный класс с тестами для всех видов менеджеров
